move cli integration tests to the reserved integration folder
also removes the need of executing `cargo run` (when runned in parallel, the tests usualy blocked each other because of this), by using `assert_cmd` library, that runs the specific profile target binary, "mocking" and also adding assertion properties to a `Command` struct.
