+++ src/test/mod.rs
@@ -14,6 +14,5 @@ mod util;
 mod bits;
 mod hasher;
 mod hvm;
-mod cli;
 mod network;
 mod node;
